added ignore to watch files

+++ src/watcher.rs
@@ -2,7 +2,9 @@ use crate::config::Service;
 use crate::kill_process::kill_process;
 use crate::spawn_service::spawn_service;
 use colored::Colorize;
+use globset::GlobSetBuilder;
 use notify::{Config, RecommendedWatcher, RecursiveMode, Watcher};
+use std::fs;
 use std::path::Path;
 use tokio::process::Child;
 use tokio::{
@@ -14,12 +16,30 @@ pub async fn run_with_watch(service: Service, color: colored::Color, max_name_le
     let (tx, mut rx) = mpsc::channel(1);
     let watch_dir = Path::new(&service.dir);
 
-    let tx_clone = tx.clone();
+    let abs_service_dir =
+        fs::canonicalize(&service.dir).unwrap_or_else(|_| Path::new(&service.dir).to_path_buf());
+    let ignore = service.clone().ignore.unwrap_or_else(Vec::new);
+
+    let mut builder = GlobSetBuilder::new();
+    for pattern in &ignore {
+        if let Ok(glob) = globset::Glob::new(pattern) {
+            builder.add(glob);
+        } else {
+            eprintln!("Invalid ignore pattern: {}", pattern);
+        }
+    }
+    let glob_set = builder.build().unwrap();
+
     let _watcher = {
         let mut watcher = RecommendedWatcher::new(
-            move |res: Result<notify::Event, notify::Error>| {
-                if res.is_ok() {
-                    let _ = tx_clone.try_send(());
+            move |res: Result<notify::Event, notify::Error>| match res {
+                Ok(event) => {
+                    if !event.paths.is_empty() {
+                        let _ = tx.try_send(event.paths.clone());
+                    }
+                }
+                Err(e) => {
+                    eprintln!("Watch error: {:?}", e);
                 }
             },
             Config::default(),
@@ -51,21 +71,37 @@ pub async fn run_with_watch(service: Service, color: colored::Color, max_name_le
 
     loop {
         tokio::select! {
-            _ = rx.recv() => {
-                if let Some(mut c) = child.take() {
-                    let padded_name = format!("{:<width$}", out_prefix.clone(), width = max_name_len);
-                    println!(
-                        "├─{} ➤ File changed, restarting service...",
-                        padded_name.color(color).bold()
-                    );
-                    kill_process(&mut c).await;
-                    sleep(Duration::from_millis(500)).await;
-                }
-                child = spawn_service(&service, color, false, max_name_len).await;
+            changed_files = rx.recv() => {
+                if let Some(paths) = changed_files {
+                    let filtered_paths: Vec<_> = paths.into_iter()
+                        .filter(|path| {
+                            let rel_path = path.strip_prefix(&abs_service_dir).unwrap_or(path);
+                            !glob_set.is_match(rel_path)
+                        })
+                        .collect();
 
-                sleep(Duration::from_millis(500)).await;
-                while rx.try_recv().is_ok() {}
+                    if filtered_paths.is_empty() {
+                        continue;
+                    }
+
+                    if let Some(mut c) = child.take() {
+                        let padded_name = format!("{:<width$}", out_prefix.clone(), width = max_name_len);
+                        println!(
+                            "├─{} ➤ File changed, restarting service...",
+                            padded_name.color(color).bold()
+                        );
+                        kill_process(&mut c).await;
+                        sleep(Duration::from_millis(500)).await;
+                    }
+
+                    child = spawn_service(&service, color, false, max_name_len).await;
+                    sleep(Duration::from_millis(500)).await;
+
+                    // drain remaining events
+                    while rx.try_recv().is_ok() {}
+                }
             }
+
             _ = async {
                 if let Some(ref mut c) = child {
                     c.wait().await
